Split queries module into appQueries and adminQueries
